backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
geographiclib==2.1
geopy==2.4.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
python-dotenv==1.1.1
typing_extensions==4.15.0
Werkzeug==3.1.3
